refactor(plotting): move combi_plot to plotting module with improved API
- Move combi_plot from SDCAnalysis to plotting.py as standalone function
- Add fontsize hierarchy: fontsize, tick_fontsize, label_fontsize, colorbar_fontsize
- Add dynamic wspace/hspace calculation based on figsize
- Add n_ticks parameter for controlling axis tick count
- Add dpi parameter with 150 default
- Support both datetime and integer indexes
- Remove auto-generated title (now opt-in)
- Move ts2/mc2 x-axis ticks to top to avoid mc1 clash
- Remove plot_two_way_sdc (replaced by combi_plot)
- Inline two_way_plot method in SDCAnalysis

Fixes combi_plot visual issues and improves modularity.
