+++ sdcpy/__init__.py
@@ -7,7 +7,7 @@ __version__ = "0.5.1"
 # Public API
 from sdcpy.core import compute_sdc, generate_correlation_map, shuffle_along_axis
 from sdcpy.io import load_from_excel, save_to_excel
-from sdcpy.plotting import plot_two_way_sdc
+from sdcpy.plotting import combi_plot
 from sdcpy.scale_dependent_correlation import SDCAnalysis
 
 __all__ = [
@@ -15,7 +15,7 @@ __all__ = [
     "compute_sdc",
     "generate_correlation_map",
     "shuffle_along_axis",
-    "plot_two_way_sdc",
+    "combi_plot",
     "save_to_excel",
     "load_from_excel",
 ]
